Correct handling of non-invertible matrix in Newton's step in enumpoly
In the Newton step in enumpoly, matrices which are singular (or close
to singular) would result in unpredictable behaviour for the
step, including in some cases failing to find an isolated equilibrium
(for example in the standard 2x2x2.nfg example, with full support).

This broadens the criteria for perturbing the initial point such that
perturbation happens if the matrix is sufficiently close
to being singular.

As part of this, the implementation of the Newton steps has been
simplified and clarified and the documentation updated, to be more
clear as to exactly how the solutions to the systems of polynomial
equations are computed.