Change to only sorting information sets on demand.
The order in which information sets, and their members, was iterated was
previously enforced to be the order encountered in a depth-first
traversal of the tree.  This scales poorly when building large games
using many operations.

This introduces a new operation, `SortInfosets` in C++ and `sort_infosets` in Python,
which does this sorting, but only on demand when called.

Sorting is done automatically in the following circumstances:
* When reading an .efg file;
* When creating the reduced strategic form of a game (to ensure consistent strategy labels);
* After every operation in the GUI (where performance is not critical relative
  to the usefulness of having lists of information sets and
  members be independent of the editing order)

Closes #483.