Address some common compiler warnings.
This implements some clean-ups that address some compiler warnings.
Principal among these are more consistent uses of size_t for values
which are related to the sizes of containers.

Along the way, where it was straightforward to do so, some passages
were re-written to avoid signed/un-signed comparisons by using
instead STL algorithms and idioms.