Make behaviour of reference counting of games consistent.
The behaviour of reference counting of games has to date not been well-defined.
An object or variable holding a reference to an individual game element (player, outcome, etc.) was
not enough to extend the lifetime of a game object if the last
user-held reference to the game went out of scope (in either C++ or Python).
In contrast, a collection object (e.g. the set of players) did hold a shared reference to
the underlying game and therefore extended the lifetime of the game.

* Standardise on the semantics that holding a reference to any component
  element or elements of a game, or another indirect reference to
  a game via e.g. a profile, extends the lifespan of the underlying
  game.  (This is the same semantics as was used in the old
  Gambit Command Language.)
* Add a section in the `pygambit` user guide briefly explaining this.

Closes #537.