Change GUI always to draw information sets fully
Previously, the GUI allowed information set drawing to be turned off entirely,
or to only be drawn for nodes at the same level.

After much reflection, these options do not seem like good ideas given
the standard way games are drawn.  Further, we hope to improve
tree layout in the near future.

Therefore, we are withdrawing this feature, and all information set connections
will be rendered.