New syntax for node children and paths (#587) (#588)
This changes the indexing operator on `Node.children` to search the labels of
actions available at the node, rather than the labels of the children nodes themselves.

It also extends the operator to accept an action to index by, the interpretation of which
is to return the node which succeeds the parent node following that action.

Game-building tutorials in the user guide have been updated to use this new feature.

Closes #587.

---------

Co-authored-by: Rahul Savani <rahul.savani@gmail.com>