Fix regression in GUI not properly numbering actions at a new information set. (#619)
Due to a previous refactoring the GUI was not generating distinct
labels at new information sets created by drag-and-drop of the
player icon.  Instead, the label "1" was assigned to all actions.

This restores the correct behaviour.  Internally, the action generation
option has been refactored to be part of the game operation itself,
in anticipation of future developments on the rules on action labels.

Closes #618.