ENH: Implement pure strategy node reachability check (#632)
This adds a property which checks whether a node is reachable by a pure strategy profile.

A node may be unreachable in a game with (an) absent-minded information set, as reaching that node
may require contradictory actions to be played if the node's history would require different actions
to be played on multiple visits to an information set.

Closes #629.