Set up boilerplate framework for adding subgames as type
This adds to C++ and Python the generic "boilerplate" code that makes
the new planned explicit representation of subgames `GameSubgameRep`
a type of `GameObject` (with its invalidation mechanism), and
the Python wrapper class `Subgame`.

A stub entry into the Python API index has also been added.

Now all that is left to do is to actually implement tests and functionality!!!!