New behaviour of `GetMinPayoff` and `GetMaxPayoff` to refer to plays of games (#655)
Previously, `GetMinPayoff` and `GetMaxPayoff` referred to the set of (non-null) outcomes in the game.  This is often not what you want when analysing a game:

* A terminal node with a null outcome is treated as a payoff of 0 and so should be taken into account;
* A contingency in the strategic game with a null outcome is treated as a payoff of 0 and so likewise;
* Outcomes at non-terminal nodes are supported; the "minimum" or "maximum" payoff in a game would be expected to be in reference to full plays of the game (cumulating payoffs).

This changes the behaviour of these functions to refer always to plays of the game, and adds new tests for {min,max}_payoff, {lp,lcp}_solve with EFGs w/ non-terminal/missing terminal outcomes.

---------

Co-authored-by: Rahul Savani <rahul.savani@gmail.com>